+++ packages/core/src/browser/addEventListener.spec.ts
@@ -1,4 +1,4 @@
-import type { Configuration } from '@datadog/browser-core'
+import type { Configuration } from '@openobserve/browser-core'
 import { stubZoneJs } from '../../test'
 import { noop } from '../tools/utils/functionUtils'
 import { addEventListener, DOM_EVENT } from './addEventListener'