+++ CHANGELOG.md
@@ -61,7 +61,7 @@
 
 - 🐛 Fix RUM slim npm package publication ([#2325](https://github.com/DataDog/browser-sdk/pull/2325))
 - 🐛 [RUM-231] Fix location.origin is "null" for file: URIs ([#2306](https://github.com/DataDog/browser-sdk/pull/2306))
-- ♻️ [RUM-232] create a `@datadog/browser-worker` package ([#2319](https://github.com/DataDog/browser-sdk/pull/2319))
+- ♻️ [RUM-232] create a `@openobserve/browser-worker` package ([#2319](https://github.com/DataDog/browser-sdk/pull/2319))
 - 🔊 Add tracekit try parse message failing telemetry ([#2322](https://github.com/DataDog/browser-sdk/pull/2322))
 - 💬 Update heavy customer data warning ([#2316](https://github.com/DataDog/browser-sdk/pull/2316))